pkg/rego_vm: wire VM in via rego plugin
Without the config, our plugin wouldn't be enabled.

Perhaps this is something we should reconsider: as it stands, the drop-in
replacement will not work: you'd have to add a configuration change for using
the rego target plugin.

Note that this was done without copying the server code. But it does require
changes to the server for rego target plugins, obviously.

Signed-off-by: Stephan Renatus <stephan@styra.com>