pkg/vm: dot of data like topdown
For better or for worse, topdown succeeds with these test cases:

    cases:
    - data:
        "2": bar
      modules:
      - |
        package test
        p := data[2]
      note: data/toplevel integer
      query: data.test.p = x
      want_result:
      - x: bar
    - data:
        nested:
          "2": bar
      modules:
      - |
        package test
        p := data.nested[2]
      note: data/nested integer
      query: data.test.p = x
      want_result:
      - x: bar
    - data:
        nested:
          "2": bar
      modules:
      - |
        package test
        p := obj[2] {
          obj := data.nested
        }
      note: "data/negative case: nested integer"
      query: data.test.p = x
      want_result: []

and with this change, load will, too.

We're keeping track of whether a target has been sourced from data via
locals.data.

It's deliberate that this flag is not carried over when the value is assigned
to another variable, because that's how topdown does it.

Signed-off-by: Stephan Renatus <stephan@styra.com>