ci: run linters (#109)
* ci: add analysis step running gcassert

Also adds a gcassert:inline directive and a more extensive comment to
pkg/vm/statements.go's noescape function.

The comment was taken from this source:
https://segment.com/blog/allocation-efficiency-in-high-performance-go-services/

We could add more analysis and linting in the future. We probably should.

While adding the CI action, I've used a manual call to `go build` for debugging.
It would not work properly unless the GOPRIVATE env var was there, together with
the git config setup. To avoid the risk of running gcassert in a way that makes
it fail silently, I've carried the vars over.

* ci: run golangci-lint

Let's try using the action instead of the Make target -- we don't have
to worry about installing the binary.

* ci: skip caches in golangci-lint

We got a bunch of these errors:

  Error: /usr/bin/tar: ../../../go/pkg/mod/golang.org/x/mod@v0.6.0-dev.0.20220419223038-86c51ed26bb4/module/module.go: Cannot open: File exists
  Error: /usr/bin/tar: ../../../go/pkg/mod/golang.org/x/mod@v0.6.0-dev.0.20220419223038-86c51ed26bb4/module/module_test.go: Cannot open: File exists
  Error: /usr/bin/tar: ../../../go/pkg/mod/golang.org/x/mod@v0.6.0-dev.0.20220419223038-86c51ed26bb4/module/pseudo.go: Cannot open: File exists
  Error: /usr/bin/tar: ../../../go/pkg/mod/golang.org/x/mod@v0.6.0-dev.0.20220419223038-86c51ed26bb4/module/pseudo_test.go: Cannot open: File exists

Disabling the caches is a blunt solution, but it should do the trick.
We can investigate better options if it becomes a burden.

Ref: https://github.com/golangci/golangci-lint-action/issues/23

Signed-off-by: Stephan Renatus <stephan@styra.com>