plugins/data: add "data" umbrella-plugin, start "kafka" plugin (#142)
* pkg/plugins: split "data" singleton plugin from "kafka" plugin

The latter can be instantiated multiple times.

There are most certainly errors in this -- especially around
reconfiguration. This is early WIP.

* start an e2e setup for the data plugin

The bitnami images are the best documented. And they even describe how
to do without ZK:

https://hub.docker.com/r/bitnami/kafka

* data/kafka: ensure we use the rego_vm; adapt FromInterface to work like ast package
* data/kafka: don't pass batch into transformation as-is

Note that we still allow for multiple return values: if it conceivable
that for one input message, there are multiple records to be written to
the store. For them, the ordering should not matter, so keeping the
set-returning rule is OK.

* data/kafka: replace manager.PrintHook with custom one

The manager's print hook depends on the request context having some
extra request-scoped metadata which we simply don't have in our use
case.

An upstream fix is reasonable, but this is the quicker path. Since a
logging overhaul for the data plugin is near anyways, this seems OK;
and much better than a runtime exception.

Signed-off-by: Stephan Renatus <stephan@styra.com>