go: rename module, adapt imports (#174)
* Update go module name in go.mod
* Update import paths in go source files

    find pkg cmd main.go -name '*.go' -exec perl -i -pe 's#inc/load/#inc/load-private/#g' {} \;)

* ko: adjust config

Following one of the steps outlined here:
https://github.com/ko-build/ko/issues/892

Also removes the KO_DOCKER_REPO from the workflow definition. The
setup-ko action doesn't need to know that, and the `deploy-ci` make
target knows it from the Makefile's export.

⚠️ The benchmark data check fails here, because the packages' names
include the module, and that has changed. I think it should resolve itself
once this is merged, or I'll have another look afterwards.

Signed-off-by: Stephan Renatus <stephan@styra.com>