liactl: add progress bar TUI (#334)
This is introducing bubbletea for the terminal UI. There are a few things we
could look into to make it nicer:

1. upload progressbar -- If the bundle used is large, or the connection is slow,
it would be nice to show something during that process. We know the file size,
so this could be quite accurate. We'd be using a io.TeeReader similar to this
example: https://github.com/charmbracelet/bubbletea/blob/de6740db2e06beb4eaf968cc7551f6d17c1b9024/examples/progress-download/main.go#L18-L40

2. spinner between finished upload and HTTP status response -- There's
some time spent when Load is parsing the bundle and enables LIA mode.
We could add some feedback there.

3.Color scheme? I've kept it the default for now.

Signed-off-by: Stephan Renatus <stephan@styra.com>