pkg/vm: Avoid computing statement position for each statement execution.
Instead, proceed through the statements sequentially, which is trivial
for most statements as they are fixed size. As a bonus, make the
instructions a bit more compact by merging type and length fields to a
single uint32.

Together with the previous commit, this brings speed up of 1-2% in
statement execution.

Signed-off-by: Teemu Koponen <koponen@styra.com>