pkg/json: Use the compact types in construction, when possible. (#505)
This way more code can benefit from the memory savings. In particular,
in particular, this makes the construction of JSON types out of golang
native JSON types.

This patch does not change all the array and object constructions to
size compact types as they can be slower to modify. That is, it
doesn't make sense to use them during the evaluation, for instance.

As a bonus, improve the comparison functions to work with binary
blobs.

Signed-off-by: Teemu Koponen <koponen@styra.com>