pkg/storage: Block accesses spanning multiple stores.
If the namespace store has multiple stores attached to it, prevent all
accesses that potentially span multiple stores: if paths 'a/b' and
'a/c' have a store attached each, one can't read from at / or
/a. Accesses to /b are accepted and will hit the default, inmem root
store. Access to a/d is not allowed as no store exists at that
location.

With a single root store, as Load stands today, this is a no-op patch
as all accesses will hit a single store anyway.

Signed-off-by: Teemu Koponen <koponen@styra.com>