data/kafka: update on policy updates; don't consume messages when without transform (#541)
Before, updates to policies would not have updated the prepared eval state 
of the kafka plugin's data transform.

This would happen if the policy wasn't there when starting (like in `load eval pol.rego`),
but was used from a bundle, or updated via the HTTP API.

The result was that without an updated transform, your consumed messages would
always be dropped.

Now, we're registering the plugins for compiler triggers -- whenever it updates, the
prepared state is refreshed. It's pretty standard procedure, the same as done with
the dropper/masker in DL, or in any server handler.

Furthermore, when the plugin is instantiated but we don't have any transform rule,
we will not consume messages. Otherwise, it would have been possible that when
using the kafka plugin with a bundle, all messages ready to be consumed would be
fetched, and dropped, before the bundle plugin was able to activate the policy that
contains the transform.

⚠️ Another thing that became apparent is that when using bundles for the transform
policies, it's critical that the bundle's data does not overlap with the data plugin's sub
tree: if it did, there would not be an error upfront, but the data plugins write attempts
would fail. (I'm about to add a bug issue for this.)

ℹ️ We're also changing the log level mapping here: When Load is configured to run with
info logging, we'll only emit warn logs of the kafka library. When Load is configured with
debug logging, we'll get all of info and *debug from the kafka library.
* plugins/data/kafka: log levels: map load<info> to kafka<warn> to reduce noise

The INFO messages of the kafka library are interesting only when
debugging. So let's not confuse users with too many logs in standard
settings.


Signed-off-by: Stephan Renatus <stephan@styra.com>