decision logs: splunk output type (#605)
This adds a new output type: splunk.

An example config is this:

output:
- type: splunk
  url: "https://http-input.foobar.hec.splunk.com/services/collector/event"
  token: opensplunkame
  batching:
    at_period: "500ms"
    compress: true # gzip payloads? 

It's meant to be configured to emit events to the "HTTP Event Collector (HEC)"

Also:
* decision_logs/tests: validate the stream YAML, ignore go leaks

Previously, we had merely been constructing structs, so the goleaks
validation was unnecessary. Now, we're also starting the stream so that
benthos' linting validates that whatever map[string]any we construct is
valid for benthos.

This made the goleaks validator fail. I've disabled it to keep going, we
can revisit the shutdown flow of benthos' inproc input (that we're
using) later.

* decision_logs: workaround data race in NewStream vs RegisterProcessor

This is odd code now. However, there's only ever one instance of a DL
stream in Load, so it's OK. But really, it's odd code.

* decision_logs: allow kafka to skip TLS validation, too

Signed-off-by: Stephan Renatus <stephan@styra.com>