cmd: Adds a license trial subcommand to create trial license keys (#653)
This commnad takes sign-up values via flags or interactively through the
terminal. It validates the values, and then requests a new trial license
key from the trail server. The new key is sent to stdout where the user
can add it as an environment variable and begin evaluationg Styra Load
immediately.

Signed-off-by: Luke Woodward <woodward.lucas@gmail.com>