cmd: Adjust missing license error messaging to encourage forward progress (#662)
This modifies the error message in cases where a license is not defined. This includes
commands like `run` as well as the `license` validation command. Output no longer goes
through the license logger, but is directly output to the stderr stream. In the
`license` command, the message is only output once instead of multiple times as
was previously the behavior.

Signed-off-by: Luke Woodward <woodward.lucas@gmail.com>