make source-examples work here (get rid of load-not-so-private) (#683)
* pkg/rego_vm: don't assume default target by default, add bjson

This is to allow using the package with "rego package" source availability.

* examples/rego: add go module for using rego package

NOTE: I had to use GONOSUMDB=github.com/srenatus/* to get
go mod tidy to work. Before, I ran into this error:

    go: github.com/open-policy-agent/opa@v0.53.1 (replaced by github.com/srenatus/opa@v0.53.2-0.20230606095611-81beeec3c92d): verifying go.mod: github.com/srenatus/opa@v0.53.2-0.20230606095611-81beeec3c92d/go.mod: reading https://sum.golang.org/lookup/github.com/srenatus/opa@v0.53.2-0.20230606095611-81beeec3c92d: 404 Not Found
	server response: not found:

The command that worked was

    GONOSUMDB=github.com/srenatus/* GOPRIVATE=github.com/styrainc go mod tidy

* pkg/{storage,plugins}: add build tags to code that requires fork
* ci/pull-request: remove gcassert (build tags)

This tool doesn't seem to be able to build with build tags, so
we cannot use at anymore at this point.

* plugins/bundle: only add ActivateOpts.Plugin with fork

This piece will likely need to be upstreamed for the SDK usage to work
but it's enough to get the basics into place before: it compiles, and Load
behaves like it did before.

* examples/sdk: scaffolding for SDK use

* build: update fork branch

Added a Store option: https://github.com/StyraInc/opa/pull/108

* pkg/plugins: comment out what isn't working/tested
* pkg/user_agent_noop: remove TODO

This feels wrong. But it's OK for now, I guess.

* test-examples-{sdk,rego}: "go mod tidy" in make target

Without this, we'd have to manually update the go.sum files of
the examples-*/ dirs with every dependency bump in top-level.

To avoid this, we'll now run the go mod tidy within the test run.

Signed-off-by: Stephan Renatus <stephan@styra.com>