Updates batching config to use a 'format' enum, deprecating the array boolean (#780)
Rather than introduce another boolean option to the S3 config which would then
ignore other boolean options, this allows batch configuration to be set to specific
formats. Out of the gate this includes 'none', 'lines', and 'array' with the
default being 'lines'. This is backwards compatible with the previous behavior.
The array flag is not removed here, rather it is marked as deprecated. All
previous configurations still function as they did previously.

Signed-off-by: Luke Woodward <woodward.lucas@gmail.com>