Introduce a preview API for performing remote preview requests (#846)
This adds an optional preview API to Enterprise OPA. Anyone with access to the OPA Rest API, either through tunneling to a remote instance or by using a running local instance, can request a decision at a specific path, with a given input. This decision will take into account any additional rego modules or data sent with the request. Sent data and rego modules are only active for the preview request and are not cached or persisted by the Enterprise OPA instance in any way. Any conflicts between the preview data and the existing data is resolved by preferring the provided preview data or policies.

This allows users to preview decisions with changes to policies and data before those changes are sent live -- within the context of the live OPA instance, and with all existing policies and data already available.