builtins: add utility libraries for vault/env usage patterns (#876)
This is using the ast.ModuleLoader apparatus, which may or may
not be a good idea -- let's find out.

One thing to wonder about is copyright: if a user builds an
optimized (i.e. partially evaluated) bundle from their rego
calling EOPA's helpers, the code for the helpers could go
into the resulting bundle. We can look into shallow inlining
controls to prohibit that; but it's a problem to worry about
later, I think.

* library: inline all vars

  So nothing is leaked when someone queries /v1/data/system/eopa/utils.
  It's somewhat more ugly, but bearable, I think.

Signed-off-by: Stephan Renatus <stephan@styra.com>