allow using rego.Function{1,2,3} with rego_vm (#901)
This looks weird -- why not just call BuiltinFuncs() ? -- but that's on purpose:
before the change made it into OPA, the function wasn't present. So we'll tip-toe
along to support the use case if we're used with a recent-enough OPA.

Through source-availability, these sources could be used with a few different OPA
versions. Most notably, the latest released version (0.55.0) doesn't have the
change we need yet.

Signed-off-by: Stephan Renatus <stephan@styra.com>