Migrates the preview API to use hooks instead of plugin OPA integration (#990)
Discussion with the team lead to the desire to enable the preview API by
default. As a plugin it would never run unless the plugin configuration
was present, but by moving it to use the hooks system, that configuration
is no longer necessary. This also adds configuration options to disable
the API, which is settable during initialization and on discovery bundle
updates.

Signed-off-by: Luke Woodward <luke@styra.com>