plugins/bundle: Deprecate entire plugin. Keep activator. (#1066)
* plugins/bundle: Deprecate entire plugin. Keep activator.

This commit fixes a discovered panic by rewiring the gRPC server to use
upstream OPA's bundle plugin for bundle info lookups. It also removes
our custom EOPA bundle plugin. The only place in the entire codebase
that needed or used it was a small sliver of code in the gRPC plugin,
which blew up with a panic when the default OPA `bundle` plugin was
appearing.

All functionality the custom plugin served appears to have been
upstreamed into mainline OPA or the OPA fork patchset.

Signed-off-by: Philip Conrad <philipaconrad@gmail.com>

* plugins/bundle: Remove more dead code.

Signed-off-by: Philip Conrad <philipaconrad@gmail.com>

---------

Signed-off-by: Philip Conrad <philipaconrad@gmail.com>