pkg/vm: Unify the Object and Set data types with json package types.
IterableObject becomes a type that does not allow any other operations
besides get and iterate. It is essentially reserved for storage
implementations implemented outside of VM package and its two
operations are used to narrow down to the data to retrieve from the
storage: whenever a VM instruction/builtin is about to execute any
other operation over the data, we are ready to (recursively) convert
the IterableObject to a JSON type (see castJSON function). Hopefully
at this point the data to retrieve and convert from storage is
specific enough the conversion to be efficient.

Previously IterableObjects provided Len and Equal too. This was
because the interface was also implemented by vm.Object and those
operations are required by statements/builtins. Now there's a
separate, more general vm.Object interface for general object type.

This paves the way to move the Object and Set types under the JSON
package. IterableObject can remain within the VM package as the JSON
types don't have to interact with the IterableObject directly.

This patch does not clean the unnecessary context passing from the
Object/Set interfaces yet.

Signed-off-by: Teemu Koponen <koponen@styra.com>