eopa_dl: avoid data race in drop/mask processor ctors (#1312)
The drop and mask processing stages we're setting up are instantiated from
`stream.Run()`. That method also blocks forever while the stream is in
operation, so we cannot do much in terms of letting it inform us. What we do
instead is add a waitgroup to the registerer, with each expected processor,
counting down in their constructors. In newStream(), we can then wait for them
all to be instantiated before returning, avoiding any data races.

Signed-off-by: Stephan Renatus <stephan@styra.com>