support datasources in eopa pull (#1335)
Downloading datasources from DAS Libraries now works as expected. In particular, this means that library code is now rooted at .styra/include/libraries rather than .styra/include, which is important so that JSON data files for the datasources or embedded in the library get loaded into the correct path in the data. hierarchy. This may require users to manually delete the .styra folder and allow eopa pull to recreate it.

EOPA-141