add regression test for combining run+datasources (#1360)
No EOPA behaviors were modified, but I have added a regression test to
ensure that the functionality required by EOPA-142 is working. It already
worked, but this will ensure we find out if it stops. The regression test
invokes syncmock and eopa pull to ensure that even if we later change
the on-disk layout, eopa pull && eopa run will continue to work with the
same config file.

EOPA-142