feat(telemetry): add bundle_sizes (#1367)
* feat(telemetry): add bundle_sizes

TODO: tests

EOPA-192

Signed-off-by: Stephan Renatus <stephan@styra.com>

* telemetry/bundle_sizes: add clusmy wait-for-first-report logic

Signed-off-by: Stephan Renatus <stephan@styra.com>

* telemetry/bundle_sizes: switch to timing-based wait

This still seems brittle, but maybe less so than before. In the previous
iteration, there was a bit of a race between `isSetup` begin _set_ and read
from the telemetry gathering goroutine.

Signed-off-by: Stephan Renatus <stephan@styra.com>

* telemetry/bundle_sizes: remove all synchronization logic, fire on USR1 signal

The rationale is that we don't need to synchronize the startup process if we
get all the bundle size data we want in the second (and subsequent) requests.

The latter is so we can test this: the normal second request comes in an hour;
we can't wait that long in  tests. So instead, we'll fire a telemetry report
whenever there's a SIGUSR1 sent. The mechanism is only activated for tests.

Signed-off-by: Stephan Renatus <stephan@styra.com>

* e2e/telemetry: add bundle_sizes tests

Signed-off-by: Stephan Renatus <stephan@styra.com>

---------

Signed-off-by: Stephan Renatus <stephan@styra.com>