telemetry: initial support for datasources (#1369)
This PR adds a new datasources field to the telemetry data that EOPA reports, which is an object with keys being datasource types, and values being the number of datasources of that type that appear in the config. There is no explicit validation, meaning that invalid datasource types in the configuration may appear in the telemetry; this was an intentional design decision, as it may reveal interesting information to us about what datasources our users _think_ EOPA supports. 

EOPA-196