keygen: retrying license stuff (#1490)
Notable parts:

* license: make Checker a struct (was interface)

The indirection wasn't useful, but it was annoying when going to a definition
and always hitting the interface def.

* license: remove mustStartBy

This complicates things, and its value isn't clear. We're now giving the
entire license verification process 48hrs, so there's no point in a fail-
safe.

This used to make sense when EKM was driving the license verification. Now,
EKM is supplying information, but the driving force is the retrier.

If EKM (or any file on disk) fails to produce a valid license within 48hrs,
we'll give up.

* license: we no longer need Strict()

Because we either start the retrier, or we don't.

* license: make synchronous check use the same code as async

A one-off check to decide if we're EOPA or OPA-fallback is the same thing
as a the retrying background check -- the only difference is that there's
only one attempt. Any kind of failure means we're going into OPA-fallback
mode.

Also contains some ctx-related cleanups.

* internal/license: use retryhttp for small-scale retry

So, besides retrying the whole dance after an hour, we'll use a retrying
http client with the keygen lib: it takes care of any networking hiccups
and, more importantly, rate limiting.

We could ignore this possibility, since the global retrier will retry
anyways, but when running `eopa run --server` with synchronous license
checking (i.e. without `--no-license-fallback`), there's only one attempt.

* internal/license: fix error wrapping

The *keygen.Error instances aren't the right thing to unwrap --
we only care for unwrapping the retryhttp errors... but those are
built from `fmt.Errorf()`, without any proper types. So this will
have to do for now.

* license: make caller decide on errors, remove  exit struct field

Also switches `eopa exec` to not do retries in the async mode: Given
that it's more like `eopa eval` with benefits, and it's very unlikely
to take longer than 48 hrs to finish, we'd otherwise practically remove
the need for a license.

EOPA-218

Signed-off-by: Stephan Renatus <stephan@styra.com>
Co-authored-by: Philip Conrad <conradp@chariot-chaser.net>