build: use .go-version to select go version (#1599)
And put 1.22.2 in it.

Before, we relied on 1.22 picking up the latest 1.22.x version via
actions/setup-go. When a new version is released, that won't appear in the
manifest used by that action right away. However, explicitly selecting that
version makes the action pull the golang binaries from Github.

I've refrained from introducing asdf, mise, aqua, or any other new tool for
this.

Signed-off-by: Stephan Renatus <stephan@styra.com>