data: add benthos input scaffolding, use for pulsar (#1676)
This includes a bit of delayed maintenance work: To use the benthos input component
for this, we had to make some changes to its code. Upstream PRs have been opened,

1. https://github.com/benthosdev/benthos/pull/2602
2. https://github.com/benthosdev/benthos/pull/2592

but we're not waiting on them to be merged, using a temporary fork instead. This will
need to be updated later.

E2E-tests aren't as comprehensive as they are for the Kafka plugin, but they exist:
testing the plain token auth and no-auth configs.

The batching for this component is hardcoded for now: not setting up manual batching
in the benthos config like we do here would result in one-element-batches being fed
into the rego transform, which is a bit inefficient. If this data plugin sees real-world use,
we might need to adjust the settings or make them configurable.

Signed-off-by: Stephan Renatus <stephan@styra.com>