capabilities: Include metadata in generated JSON
In https://github.com/StyraInc/regal/issues/1218, we are interested in
showing the correct content on hover in VS Code for the Enterprise OPA
built-ins.

Following:
- https://github.com/StyraInc/enterprise-opa-private/pull/2060
- https://github.com/StyraInc/enterprise-opa-private/pull/2061

We now need to get this data into the public repo for Regal to embed.

https://github.com/StyraInc/enterprise-opa/tree/main/capabilities

Which is done via:

https://github.com/StyraInc/enterprise-opa-private/blob/main/.github/workflows/push-tags.yaml#L166

When the files are in the regal repo:

https://github.com/StyraInc/regal/tree/main/internal/capabilities/embedded/eopa

They are loaded here:

https://github.com/StyraInc/regal/blob/main/internal/capabilities/capabilities.go#L182-L187

And when this happens the data is needed to show in the hover tool tips.

STY-0