pkg/compile: support internal.member_2
This is now possible:

    include if input.fruits.name in {"apple", "cherry", "pineapple"}

While these still aren't:

    include if "apple" in input.basket.elements

    include if "key", input.fruits.name in {"key": "apple", "o": "cherry"}

E2E tests tell us that they are fine with the SQL output: PG/MySQL/SQLServer
understand the WHERE clauses we emit. However, when emitting UCAST, we'll
probably need to emit errors, as Prisma/LINQ might not be able to deal with
them.

Signed-off-by: Stephan Renatus <stephan@styra.com>