pkg/plugins/bundle/activator: Auto-detect v0/v1 for bundles. (#2212)
This commit fixes a bug that has been present since at least OPA
v1.31.0, and possibly since v1.30.1, which caused bundles with
`rego_version: 0` to be handled incorrectly (as Rego v1 bundles, instead
of as Rego v0 bundles). This caused bundles to unexpectedly fail, unless
the user passed in the `--v0-compatible` flag, which was not the
intended behavior.

The solution was adding an ugly-but-effective workaround, which allows
our custom bundle activator to correctly load and unload bundles,
regardless of whether the bundle had a `rego_version` field set for
Rego v0 or v1.

The logic to make that happen requires dredging up the original
`rego_version` of every bundle, and checking to see if any require v0.
If so, it will retry parsing during bundle unload, using the RegoV0
parsing option. If things still fail, an error is returned, as normal.
Also, if the policy parsed okay with v0, but the bundle requires v1, we
emit an error there too.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>
Signed-off-by: Stephan Renatus <stephan@styra.com>
Co-authored-by: Stephan Renatus <stephan@styra.com>