builtins: remove ucast.expand, ucast.as_sql
They are no longer used from any of our demo setups, and with the Compile API
extensions, we no longer need to construct UCAST in Rego. There is thus no
need to expand or translate it.

Alternatively, we could have exposed the Compile API to Rego -- but if we do
that (we still can), we'll have _those builtins_ return UCAST or SQL, same as
the Compile API. So there's no need for these builtins anymore.

Signed-off-by: Stephan Renatus <stephan@styra.com>