spike: implement 'rego.compile()' (#2339)
* data filtering: implement 'rego.compile()'

* library: add test filter helper to EOPA's library

Also make it only DROP the tables if it's using :memory:. Just because if you are
debugging your tests, you might appreciate having the tables around for inspection.

* rego.compile(): expose error handling via 'raise_error' configurable

Same as in sql.send, it defaults to `true`. If you set it to `false`, certain
translation-related errors are returned instead of turning the result into
undefined.

For `eopa test`, it doesn't matter as much, any errors returned by builtins
are printed.

Signed-off-by: Stephan Renatus <stephan@styra.com>