pkg/plugins/data/localfile: Add local file polling data source plugin. (#2365)
This commit adds a data source plugin for reading in local
files on disk. It runs a polling loop, and checks the hash of the file
contents against the last known hash. If there is a diff, it re-parses
the file contents, transforms them, and then re-ingests the results.

* pkg/plugins/localfile: Add tests + better file path check.
* pkg/plugins/data: plumb in the new plugin.
* pkg/plugins/data/localfile: Add back "*Owned" test.
* pkg/plugins/data/localfile: Remove dead constants.
* plugins/data/localfile: Remove Timeout parameter.

It's unused currently, and would be tricky to add. The main use case
would be detected and aborting from slow I/O operations. For now, this
is a non-issue. However, if users try fetching files from NFS's, I/O
overheads may actually become significant.

* plugins/data/localfile: Add YAML test.
* pkg/plugins/data/localfile: Add filetype-from-ext + config key.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>