pkg/compile: Add support for mask evals. (#2438)
This commit adds a new feature to the extended /v1/compile API: column
masks from evaluating a mask rule.

This commit involves extensions or modifications to:
 - Request Format: New option parameter: `mask_rule`.
 - Response Format: New optional field: `Result.Masks`.
 - Caches: New internal cache for looking up the masking rule name.
 - Metrics:
   - New timer metric for mask rule eval.
   - Splits metrics for `unknowns` and `mask` paths.
 - Decision Logs: New custom field for recording the mask rule name.

On the error path for mask rule lookups, we also do a Levenshtein fuzzy
string search across *all* rules present on the compiler, which may
prove expensive for OPAs with a large number of rules loaded. The
improvement in the debugging experience makes it worthwhile, however.

See: EOPA-404

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>