pkg/compile: Add package-local mask rule name matching. (#2501)
If the user provides a `mask_rule` annotation (or JSON option)
specifying just a bare name, like `masks`, we now try to match it
against something in the same package as the rule.

Example:

    # METADATA
    # scope: package
    # custom:
    #   mask_rule: masks
    package filters

    default include := true

    masks.tickets.description.replace.value := "***" if {
        "reader" in data.roles[input.tenant][input.user]
    }

If the user has any other kind of ref that doesn't start with
`data.<whatever>`, they'll get the usual fuzzy-matching behavior.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>