compile: re-do compiler+annotation handling (#2510)
Parsing the annotations "on demand" and replacing the compiler
afterwards is a bit too brittle to get right. Now, we're instead
dealing with annotations parsing right away, on init, and whenever a
policy has changed. The cloned compiler is then used for searching
through the annotations set for unknowns and mask rules.

Signed-off-by: Stephan Renatus <stephan@styra.com>