vm/rego.compile: make it possible to use this with OSS OPA
The flow is slightly different: when called in the non-fork setup (like
in styra-playground), it expects the modules to be part of the call
options passed.

As it happens, that's what the styra-playground does anyways: rego.eval
and rego.compile now behave more alike.

For EOPA, nothing changes (yet) with this diff: it'll work as it did
before. It's only SDK-like usage (via source-available builds) where it
makes a difference.

Signed-off-by: Stephan Renatus <stephan@styra.com>