feat: EOPA becomes licensed when discovery detects a DAS bundle. (#2653)
This commit introduces a new licensing quirk: when discovery detects a
discovery bundle from a Styra DAS instance, it marks the EOPA instance
as licensed.

Because the security requirements for this feature are quite lax, the
distinguishing feature for DAS-specific discovery bundles is just a
"magic field", of the form:

    {
        ... other fields ...,
        "eopa": {
            "license": "VALID"
        }
    }

To detect discovery bundles of this form, we have to first make sure the
config is doing discovery at all, and if so, we then fire up an
isolated, one-off discovery plugin instance to fetch the bundle down. If
the magic field is present, we're licensed. If not, we're not.

Note: This can result in a "double-fetch" discovery scenario, unless
discovery bundle persistence is turned on. In that case, the second
"real" discovery plugin will fetch the fresh discovery bundle out of the
disk cache.

New CLI flags for the `run` command:
 - `--license-discovery-timeout` :: Discovery-based license check timeout
   in seconds. (Default: 30 seconds)
 - `--no-discovery-license-check` :: Boolean flag to disable
   discovery-based license check. (Default: false)

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>