pkg/plugins/data/s3: Simplify resolver implementation.
This commit is what we actually wanted-- I was misled earlier by LLMs
into thinking that the `Endpoint` field no longer existed *anywhere*.
However, when exploring the `s3.EndpointParameter` type, I discovered an
exact 1:1 match for the old `Endpoint` field.

The fix is to only set that field in the custom endpoint resolver, and
now everything works more or less like it did before.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>