chore: Upgrade to AWS SDK V2 (#61)
* pkg/builtins/dynamodb: Upgrade to AWS SDK v2.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>

* pkg/plugins/data/s3: Upgrade to AWS SDK v2.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>

* go.mod: Bump version deps.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>

* pkg/builtins/dynamodb: Appease linter.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>

* pkg/plugins/data/s3: Simplify resolver implementation.

I had thought that the `Endpoint` field no longer existed *anywhere*.
However, when exploring the `s3.EndpointParameter` type, I discovered an
exact 1:1 match for the old `Endpoint` field.

The fix is to only set that field in the custom endpoint resolver, and
now everything works more or less like it did before.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>

---------

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>