pkg: Fix deprecation warnings from upstream OPA. (#265)
This commit fixes two deprecation warnings that occur
after upgrading the OPA version for the project.

These deprecations are:
 - v1/server/writer/writer.go: `writer.JSON()`
 - v1/bundle/store.go: `bundle.LegacyReadRevisionFromStore()`

For the `writer.JSON` call, I pulled in the code
from upstream, because we do need the efficiency of that
implementation for Batch Queries, and we can generally
ensure that the encoding error case is not going to happen.

For the `bundle.LegacyReadRevisionFromStore` call,
it's removed from all live codepaths in upstream OPA,
so we should be good to remove it here as well.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>